Split massive class. Also don't rescan if player didn't move
